修改门店接口

+++ myedu/src/main/java/com/myedu/app/store/APPYunCourseController.java
@@ -1,10 +1,14 @@
 package com.myedu.app.store;
 
+import com.myedu.common.utils.DateUtils;
 import com.myedu.common.utils.SecurityUtils;
+import com.myedu.common.utils.ServletUtils;
 import com.myedu.common.utils.StringUtils;
 import com.myedu.common.utils.poi.ExcelUtil;
 import com.myedu.framework.aspectj.lang.annotation.Log;
 import com.myedu.framework.aspectj.lang.enums.BusinessType;
+import com.myedu.framework.security.LoginUser;
+import com.myedu.framework.security.service.TokenService;
 import com.myedu.framework.web.controller.BaseController;
 import com.myedu.framework.web.domain.AjaxResult;
 import com.myedu.framework.web.page.TableDataInfo;
@@ -21,6 +25,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -43,19 +48,25 @@ public class APPYunCourseController extends BaseController
     private ISysGradeService sysGradeService;
     @Autowired
     private IYunStoreService yunStoreService;
+    @Autowired
+    private TokenService tokenService;
     /**
      * 查询课程列表
      */
     @ApiOperation("查询课程列表")
     @ApiImplicitParam(name = "yunCourse", value = "查询课程列表",
             dataType = "YunCourse")
-    @PreAuthorize("@ss.hasPermi('store:course:list')")
     @GetMapping("/list")
     public TableDataInfo list(YunCourse yunCourse)
     {
-        startPage();
-        List<YunCourseVo> list = yunCourseService.selectYunCourseList(yunCourse);
-        return getDataTable(list);
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            startPage();
+            List<YunCourseVo> list = yunCourseService.selectYunCourseList(yunCourse);
+            return getDataTable(list);
+        }else{
+            return getDataTableLose(null);
+        }
     }
 
     /**
@@ -64,7 +75,6 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("导出课程列表")
     @ApiImplicitParam(name = "yunCourse", value = "导出课程列表",
             dataType = "YunCourse")
-    @PreAuthorize("@ss.hasPermi('store:course:export')")
     @Log(title = "课程", businessType = BusinessType.EXPORT)
     @GetMapping("/export")
     public AjaxResult export(YunCourse yunCourse)
@@ -79,23 +89,26 @@ public class APPYunCourseController extends BaseController
      */
     @ApiOperation("获取课程详细信息")
     @ApiImplicitParam(name = "id", value = "获取课程详细信息",
-            dataType = "Long")
-    @PreAuthorize("@ss.hasPermi('store:course:query')")
-    @GetMapping(value = { "/", "/{id}" })
-    public AjaxResult getInfo(@PathVariable(value = "id", required = false) Long id)
+            dataType = "Long", required = true, paramType = "path")
+    @GetMapping(value = { "/", "/{id}" }, produces = MediaType.APPLICATION_JSON_VALUE)
+    public AjaxResult getInfo(@PathVariable Long id)
     {
-        AjaxResult ajax = AjaxResult.success();
-        SysGrade sysGrade=new SysGrade();
-        ajax.put("sysGrades", sysGradeService.selectSysGradeList(sysGrade));
-        YunStoreVo yunStore=new YunStoreVo();
-        yunStore.setCreateById(SecurityUtils.getUserId());
-        List<YunStoreVo> stores=yunStoreService.selectYunStoreList(yunStore);
-        ajax.put("stores", stores);
-        if (StringUtils.isNotNull(id))
-        {
-            ajax.put(AjaxResult.DATA_TAG,yunCourseService.selectYunCourseById(id));
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if(loginUser!=null) {
+            AjaxResult ajax = AjaxResult.success();
+            SysGrade sysGrade = new SysGrade();
+            ajax.put("sysGrades", sysGradeService.selectSysGradeList(sysGrade));
+            YunStoreVo yunStore = new YunStoreVo();
+            yunStore.setCreateById(SecurityUtils.getUserId());
+            List<YunStoreVo> stores = yunStoreService.selectYunStoreList(yunStore);
+            ajax.put("stores", stores);
+            if (StringUtils.isNotNull(id)) {
+                ajax.put(AjaxResult.DATA_TAG, yunCourseService.selectYunCourseById(id));
+            }
+            return ajax;
+        }else{
+            return AjaxResult.error("token无效");
         }
-        return ajax;
     }
 
     /**
@@ -104,12 +117,19 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("新增课程")
     @ApiImplicitParam(name = "yunCourse", value = "新增课程",
             dataType = "YunCourse")
-    @PreAuthorize("@ss.hasPermi('store:course:add')")
     @Log(title = "课程", businessType = BusinessType.INSERT)
     @PostMapping
-    public AjaxResult add(@RequestBody YunCourse yunCourse)
+    public AjaxResult add(YunCourse yunCourse)
     {
-        return toAjax(yunCourseService.insertYunCourse(yunCourse));
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            yunCourse.setCreateById(SecurityUtils.getUserId());
+            yunCourse.setCreateBy(SecurityUtils.getUsername());
+            yunCourse.setCreateTime(DateUtils.getNowDate());
+            return toAjax(yunCourseService.insertYunCourse(yunCourse));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -118,12 +138,17 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("修改课程")
     @ApiImplicitParam(name = "yunCourse", value = "修改课程",
             dataType = "YunCourse")
-    @PreAuthorize("@ss.hasPermi('store:course:edit')")
     @Log(title = "课程", businessType = BusinessType.UPDATE)
     @PutMapping
-    public AjaxResult edit(@RequestBody YunCourse yunCourse)
+    public AjaxResult edit(YunCourse yunCourse)
     {
-        return toAjax(yunCourseService.updateYunCourse(yunCourse));
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            yunCourse.setUpdateTime(DateUtils.getNowDate());
+            return toAjax(yunCourseService.updateYunCourse(yunCourse));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -132,12 +157,16 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("删除课程")
     @ApiImplicitParam(name = "ids", value = "删除课程",
             dataType = "Long[]")
-    @PreAuthorize("@ss.hasPermi('store:course:remove')")
     @Log(title = "课程", businessType = BusinessType.DELETE)
 	@DeleteMapping("/{ids}")
     public AjaxResult remove(@PathVariable Long[] ids)
     {
-        return toAjax(yunCourseService.deleteYunCourseByIds(ids));
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            return toAjax(yunCourseService.deleteYunCourseByIds(ids));
+        }else{
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -148,20 +177,24 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("更改课程状态下线")
     @ApiImplicitParam(name = "ids", value = "更改课程状态下线",
             dataType = "Long[]")
-    @PreAuthorize("@ss.hasPermi('store:course:changeStatusOff')")
     @Log(title = "课程", businessType = BusinessType.UPDATE)
     @GetMapping("/changeStatusOff/{ids}")
     public AjaxResult changeStatusOff(@PathVariable Long[] ids)
     {
-        int rows=0;
-        for (Long id:ids) {
-            YunCourse yunCourse= yunCourseService.selectYunCourseById(id);
-            if(yunCourse!=null){
-                yunCourse.setStatus(CourseType.OFFLINE.getCode());
-                rows=yunCourseService.updateYunCourse(yunCourse);
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            int rows=0;
+            for (Long id:ids) {
+                YunCourse yunCourse= yunCourseService.selectYunCourseById(id);
+                if(yunCourse!=null){
+                    yunCourse.setStatus(CourseType.OFFLINE.getCode());
+                    rows=yunCourseService.updateYunCourse(yunCourse);
+                }
             }
-        }
-        return toAjax(rows);
+            return toAjax(rows);
+            }else{
+                return AjaxResult.error("token无效");
+            }
     }
 
     /**
@@ -172,11 +205,12 @@ public class APPYunCourseController extends BaseController
     @ApiOperation("更改课程状态在售")
     @ApiImplicitParam(name = "ids", value = "更改课程状态在售",
             dataType = "Long[]")
-    @PreAuthorize("@ss.hasPermi('store:course:changeStatusOn')")
     @Log(title = "课程", businessType = BusinessType.UPDATE)
     @GetMapping("/changeStatusOn/{ids}")
     public AjaxResult changeStatusOn(@PathVariable Long[] ids)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
         int rows=0;
         for (Long id:ids) {
             YunCourse yunCourse= yunCourseService.selectYunCourseById(id);
@@ -186,5 +220,8 @@ public class APPYunCourseController extends BaseController
             }
         }
         return toAjax(rows);
+        }else{
+            return AjaxResult.error("token无效");
+        }
     }
 }
