+++ myedu/src/main/java/com/myedu/app/store/APPYunStoreCouponController.java
@@ -1,8 +1,13 @@
 package com.myedu.app.store;
 
+import com.myedu.common.utils.DateUtils;
+import com.myedu.common.utils.SecurityUtils;
+import com.myedu.common.utils.ServletUtils;
 import com.myedu.common.utils.poi.ExcelUtil;
 import com.myedu.framework.aspectj.lang.annotation.Log;
 import com.myedu.framework.aspectj.lang.enums.BusinessType;
+import com.myedu.framework.security.LoginUser;
+import com.myedu.framework.security.service.TokenService;
 import com.myedu.framework.web.controller.BaseController;
 import com.myedu.framework.web.domain.AjaxResult;
 import com.myedu.framework.web.page.TableDataInfo;
@@ -14,6 +19,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -34,20 +40,25 @@ public class APPYunStoreCouponController extends BaseController
     private IYunStoreCouponService yunStoreCouponService;
     @Autowired
     private IYunStoreCouponIssueService yunStoreCouponIssueService;
-
+    @Autowired
+    private TokenService tokenService;
     /**
      * 查询店铺优惠券列表
      */
     @ApiOperation("查询店铺优惠券列表")
     @ApiImplicitParam(name = "yunStoreCoupon", value = "查询店铺优惠券列表",
             dataType = "YunStoreCoupon")
-    @PreAuthorize("@ss.hasPermi('store:coupon:list')")
     @GetMapping("/list")
     public TableDataInfo list(YunStoreCoupon yunStoreCoupon)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
         startPage();
         List<YunStoreCoupon> list = yunStoreCouponService.selectYunStoreCouponList(yunStoreCoupon);
         return getDataTable(list);
+        }else{
+            return getDataTableLose(null);
+        }
     }
 
     /**
@@ -56,7 +67,6 @@ public class APPYunStoreCouponController extends BaseController
     @ApiOperation("导出店铺优惠券列表")
     @ApiImplicitParam(name = "yunStoreCoupon", value = "导出店铺优惠券列表",
             dataType = "YunStoreCoupon")
-    @PreAuthorize("@ss.hasPermi('store:coupon:export')")
     @Log(title = "店铺优惠券", businessType = BusinessType.EXPORT)
     @GetMapping("/export")
     public AjaxResult export(YunStoreCoupon yunStoreCoupon)
@@ -71,12 +81,16 @@ public class APPYunStoreCouponController extends BaseController
      */
     @ApiOperation("导出店铺优惠券列表")
     @ApiImplicitParam(name = "id", value = "导出店铺优惠券列表",
-            dataType = "Long")
-    @PreAuthorize("@ss.hasPermi('store:coupon:query')")
-    @GetMapping(value = "/{id}")
-    public AjaxResult getInfo(@PathVariable("id") Long id)
+            dataType = "Long", required = true, paramType = "path")
+    @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public AjaxResult getInfo(@PathVariable Long id)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if(loginUser!=null) {
         return AjaxResult.success(yunStoreCouponService.selectYunStoreCouponById(id));
+        }else{
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -85,12 +99,19 @@ public class APPYunStoreCouponController extends BaseController
     @ApiOperation("新增店铺优惠券")
     @ApiImplicitParam(name = "yunStoreCoupon", value = "新增店铺优惠券",
             dataType = "YunStoreCoupon")
-    @PreAuthorize("@ss.hasPermi('store:coupon:add')")
     @Log(title = "店铺优惠券", businessType = BusinessType.INSERT)
     @PostMapping
     public AjaxResult add(@RequestBody YunStoreCoupon yunStoreCoupon)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            yunStoreCoupon.setCreateById(loginUser.getUser().getUserId());
+            yunStoreCoupon.setCreateBy(loginUser.getUser().getNickName());
+            yunStoreCoupon.setCreateTime(DateUtils.getNowDate());
         return toAjax(yunStoreCouponService.insertYunStoreCoupon(yunStoreCoupon));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -99,12 +120,17 @@ public class APPYunStoreCouponController extends BaseController
     @ApiOperation("修改店铺优惠券")
     @ApiImplicitParam(name = "yunStoreCoupon", value = "修改店铺优惠券",
             dataType = "YunStoreCoupon")
-    @PreAuthorize("@ss.hasPermi('store:coupon:edit')")
     @Log(title = "店铺优惠券", businessType = BusinessType.UPDATE)
     @PutMapping
     public AjaxResult edit(@RequestBody YunStoreCoupon yunStoreCoupon)
     {
-        return toAjax(yunStoreCouponService.updateYunStoreCoupon(yunStoreCoupon));
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
+            yunStoreCoupon.setUpdateTime(DateUtils.getNowDate());
+            return toAjax(yunStoreCouponService.updateYunStoreCoupon(yunStoreCoupon));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -113,12 +139,16 @@ public class APPYunStoreCouponController extends BaseController
     @ApiOperation("删除店铺优惠券")
     @ApiImplicitParam(name = "ids", value = "删除店铺优惠券",
             dataType = "Long[]")
-    @PreAuthorize("@ss.hasPermi('store:coupon:remove')")
     @Log(title = "店铺优惠券", businessType = BusinessType.DELETE)
 	@DeleteMapping("/{ids}")
     public AjaxResult remove(@PathVariable Long[] ids)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
         return toAjax(yunStoreCouponService.deleteYunStoreCouponByIds(ids));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 
     /**
@@ -127,15 +157,19 @@ public class APPYunStoreCouponController extends BaseController
     @ApiOperation("发布店铺优惠券")
     @ApiImplicitParam(name = "yunStoreCouponIssue", value = "发布店铺优惠券",
             dataType = "YunStoreCouponIssue")
-    @PreAuthorize("@ss.hasPermi('store:coupon:publish')")
     @Log(title = "店铺优惠券", businessType = BusinessType.INSERT)
     @PostMapping("/publish")
     public AjaxResult publish(@RequestBody YunStoreCouponIssue yunStoreCouponIssue)
     {
+        LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+        if (loginUser!=null) {
         Long couponId=yunStoreCouponIssue.getCid();
         YunStoreCoupon yunStoreCoupon=yunStoreCouponService.selectYunStoreCouponById(couponId);
         yunStoreCoupon.setPublishNum(yunStoreCoupon.getPublishNum()+1);
         yunStoreCouponService.updateYunStoreCoupon(yunStoreCoupon);
         return toAjax(yunStoreCouponIssueService.insertYunStoreCouponIssue(yunStoreCouponIssue));
+        }else {
+            return AjaxResult.error("token无效");
+        }
     }
 }
